// cypress/types/user.types.ts
export interface LoginCredentials {
  loginName: string;
  password: string;
}

export interface RegistrationData {
  firstName: string;
  lastName: string;
  email: string;
  telephone: string;
  address1: string;
  city: string;
  country: string;
  zone: string;
  postcode: string;
  loginName: string;
  password: string;
  confirmPassword: string;
}